+++ packages/visual-editor/src/components/Dropzone/useComponentProps.ts
@@ -35,6 +35,14 @@ import { HYPERLINK_DEFAULT_PATTERN } from '@contentful/experiences-core/constant
 
 type ComponentProps = StyleProps | Record<string, PrimitiveValue | Link<'Entry'> | Link<'Asset'>>;
 
+type ResolvedComponentProps = ComponentProps & {
+  children?: React.JSX.Element | undefined;
+  className: string;
+  editorMode: boolean;
+  node: ExperienceTreeNode;
+  renderDropzone: RenderDropzoneFunction;
+};
+
 type UseComponentProps = {
   node: ExperienceTreeNode;
   resolveDesignValue: ResolveDesignValueType;
@@ -161,6 +169,13 @@ export const useComponentProps = ({
             ...acc,
             [variableName]: value,
           };
+        } else if (variableMapping.type === 'ComponentValue') {
+          // We are rendering a pattern (assembly) entry. Content properties cannot be edited in this,
+          // so we always render the default value
+          return {
+            ...acc,
+            [variableName]: variableDefinition.defaultValue,
+          };
         } else {
           return { ...acc };
         }
@@ -239,7 +254,7 @@ export const useComponentProps = ({
   const stylesToKeep = ['cfImageAsset'];
   const stylesToRemove = CF_STYLE_ATTRIBUTES.filter((style) => !stylesToKeep.includes(style));
 
-  const componentProps = {
+  const componentProps: ResolvedComponentProps = {
     className: props.cfSsrClassName ?? componentClass,
     editorMode: true,
     node,